// src/app/models/shorten.ts
export interface DataUrl {
  code: string;
  short_link: string;
  full_short_link: string;
  short_link2: string;
  full_short_link2: string;
  share_link: string;
  full_share_link: string;
  original_link: string;
  copied: boolean;
}

export interface DataResponse {
  ok: boolean;
  result: DataUrl;
}

export interface DynamicModalData {
  title: string;
  message: string;
}